# ondexia.contracts/openapi.yaml
...
openapi: 3.0.1
info:
  title: Ondexia — API Core
  description: |-
    Nucleo comercial: almacen, compras, ventas y configuracion.

    Esta API no habla con SUNAT. La emision de comprobantes se publica en una cola y la atiende ondexia-facturacion (DTE §3.3).
  version: 0.1.0
servers:
- url: http://localhost
  description: Generated server url
security:
- jwt: []
...
paths:
  /api/v1/ventas/clientes/{id}:
    get:
      tags:
      - Clientes
      summary: Ficha de un cliente
      operationId: obtener
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaCliente"
    put:
      tags:
      - Clientes
      summary: Actualiza contacto y nombre. El documento no cambia
      operationId: actualizar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEdicion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaCliente"
  /api/v1/ventas/clientes/{id}/estado:
    put:
      tags:
      - Clientes
      summary: "Activa o desactiva. Nunca se borra: los comprobantes lo referencian"
      operationId: cambiarEstado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaCliente"
...
  /api/v1/ventas/cajas/{id}/estado:
    put:
      tags:
      - Cajas
      summary: Activa o desactiva una caja
      description: "No se desactiva con una sesión abierta. Nunca se borra: sus sesiones\
        \ la referencian."
      operationId: cambiarEstado_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaCaja"
...
  /api/v1/perfil:
    get:
      tags:
      - Perfil
      summary: Mis datos
      operationId: ver
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DatosDePerfil"
    put:
      tags:
      - Perfil
      summary: Actualiza mi nombre y mi teléfono
      description: "El correo no se puede cambiar: es la credencial de Cognito, y\
        \ cambiarlo aquí dejaría la fila apuntando a un buzón con el que ya no se\
        \ puede entrar. La contraseña tampoco pasa por la API — vive en Cognito."
      operationId: actualizar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionPerfil"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DatosDePerfil"
...
  /api/v1/configuracion/usuarios/{asignacionId}/estado:
    put:
      tags:
      - Usuarios
      summary: Activa o desactiva a la persona
      description: "Afecta a toda la cuenta, no solo a esta empresa: es la única forma\
        \ de revocar el acceso antes de que caduque su token."
      operationId: cambiarEstado_2
      parameters:
      - name: asignacionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaUsuario"
  /api/v1/configuracion/series/{id}/estado:
    put:
      tags:
      - Series
      summary: Activa o desactiva una serie
      description: "Lo único editable. Desactivar no borra ni reinicia: el último\
        \ número queda como estaba, de modo que reactivarla continúa donde se quedó\
        ."
      operationId: cambiarEstado_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaSerie"
...
  /api/v1/configuracion/establecimientos/{id}:
    put:
      tags:
      - Establecimientos
      summary: Actualiza un establecimiento
      description: "El código no se puede cambiar: identifica al anexo ante SUNAT."
      operationId: actualizar_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEdicion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEstablecimiento"
  /api/v1/configuracion/establecimientos/{id}/estado:
    put:
      tags:
      - Establecimientos
      summary: Activa o desactiva un establecimiento
      description: "Nunca borra. Un establecimiento aparece en los comprobantes ya\
        \ emitidos: borrarlo dejaria documentos apuntando a nada. Desactivarlo lo\
        \ saca de los desplegables y conserva su codigo, sus series y su numeracion,\
        \ de modo que reactivarlo lo devuelve tal como estaba."
      operationId: cambiarEstado_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEstablecimiento"
  /api/v1/configuracion/empresa:
    get:
      tags:
      - Empresa
      summary: Datos de la empresa activa
      operationId: consultar
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEmpresa"
    put:
      tags:
      - Empresa
      summary: Actualiza lo editable de la empresa
      description: "Solo el nombre comercial y la cuenta de detracciones, que son\
        \ los unicos campos que no vienen de SUNAT. La razon social, el domicilio\
        \ y el ubigeo no estan en el cuerpo: un valor distinto del padron hace que\
        \ SUNAT rechace todos los comprobantes de la empresa. Para traerlos del padron\
        \ esta POST /empresa/verificacion."
      operationId: actualizar_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEmpresa"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEmpresa"
...
  /api/v1/configuracion/comprobantes/{codigo}:
    put:
      tags:
      - Comprobantes
      summary: Habilita o deshabilita un tipo
      description: "No se puede deshabilitar un tipo con series activas: la pantalla\
        \ diría una cosa y el sistema haría otra. Responde 400 con `tipo_con_series_activas`."
      operationId: cambiarEstado_5
      parameters:
      - name: codigo
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaTipo"
  /api/v1/almacen/productos/{id}:
    get:
      tags:
      - Productos
      summary: Ficha de un producto
      operationId: obtener_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaProducto"
    put:
      tags:
      - Productos
      summary: Actualiza la ficha. El código no cambia
      operationId: actualizar_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEdicionProducto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaProducto"
...
  /api/v1/almacen/productos/{id}/estado:
    put:
      tags:
      - Productos
      summary: "Activa o desactiva. Nunca se borra: los comprobantes lo referencian"
      operationId: cambiarEstado_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaProducto"
  /api/v1/almacen/almacenes/{id}:
    put:
      tags:
      - Almacenes
      summary: Actualiza un almacén
      description: "El código no cambia: identifica al almacén en los movimientos\
        \ ya registrados."
      operationId: actualizar_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEdicion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaAlmacen"
  /api/v1/almacen/almacenes/{id}/estado:
    put:
      tags:
      - Almacenes
      summary: Activa o desactiva un almacen
      description: "Nunca borra: aparece en cada movimiento de stock que lo toco."
      operationId: cambiarEstado_7
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionEstado"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaAlmacen"
...
  /api/v1/ventas/comprobantes:
    get:
      tags:
      - Ventas
      summary: "Los últimos comprobantes; con `tipo`, solo de ese tipo"
      operationId: comprobantes
      parameters:
      - name: tipo
        in: query
        required: false
        schema:
          type: string
          enum:
          - FACTURA
          - BOLETA
          - NOTA_CREDITO
          - NOTA_DEBITO
          - GUIA_REMISION
          - NOTA_VENTA
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaResumen"
    post:
      tags:
      - Ventas
      summary: Registra una boleta o una factura
      description: "Queda PENDIENTE y sale hacia SUNAT al confirmar. `tipo`: BOLETA\
        \ o FACTURA. Responde 400 con `emision_no_configurada` si la empresa no cargó\
        \ certificado y clave SOL. El estado ante SUNAT se consulta en /comprobantes/{id}/sunat."
      operationId: emitirComprobante
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionComprobante"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEmision"
...
  /api/v1/ventas/clientes:
    get:
      tags:
      - Clientes
      summary: Lista o busca clientes
      description: "Sin `q`, todos. Con `q`, por nombre o documento, hasta 50."
      operationId: listar
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaCliente"
    post:
      tags:
      - Clientes
      summary: Registra un cliente
      description: "Con RUC puede traer la atestación de la consulta al padrón: la\
        \ razón social y el domicilio salen de ella."
      operationId: registrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionNuevo"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaCliente"
  /api/v1/ventas/clientes/{id}/verificacion:
    post:
      tags:
      - Clientes
      summary: Vuelve a verificar el RUC contra el padrón con una atestación nueva
      operationId: verificar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionVerificacion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaCliente"
  /api/v1/ventas/cajas:
    get:
      tags:
      - Cajas
      summary: "Lista las cajas que el usuario alcanza, con su sesión abierta si la\
        \ hay"
      operationId: listar_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaCaja"
    post:
      tags:
      - Cajas
      summary: Registra una caja en un establecimiento
      description: El código es único por establecimiento y no cambia después.
      operationId: registrar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionNueva"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaCaja"
...
  /api/v1/registro:
    post:
      tags:
      - Registro
      summary: Da de alta la cuenta del usuario autenticado
      description: "Crea cuenta, usuario administrador, empresa y el establecimiento\
        \ de casa matriz (0000) en una sola transacción. La identidad sale del token,\
        \ nunca del cuerpo: aceptarla como parámetro permitiría darse de alta suplantando\
        \ a otra persona. Los datos de la empresa salen de la atestación firmada que\
        \ devuelve GET /consultas/ruc/{ruc}, así que no hay alta posible con un RUC\
        \ que no esté ACTIVO y HABIDO en SUNAT. Responde 400 si la atestación no es\
        \ válida o caducó, y 409 si el RUC ya está registrado."
      operationId: registrar_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionRegistro"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaRegistro"
...
  /api/v1/configuracion/series:
    get:
      tags:
      - Series
      summary: "Lista las series, activas e inactivas"
      operationId: listar_3
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaSerie"
    post:
      tags:
      - Series
      summary: Da de alta una serie
      description: "La primera letra la fija SUNAT según el tipo: F para factura,\
        \ B para boleta, T para guía. El `numeroInicial` es para quien migra desde\
        \ otro sistema y no puede volver a numerar desde el 1; se fija aquí y no se\
        \ vuelve a tocar."
      operationId: registrar_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionNueva"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaSerie"
...
  /api/v1/configuracion/establecimientos:
    get:
      tags:
      - Establecimientos
      summary: "Lista los establecimientos, activos e inactivos"
      operationId: listar_4
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaEstablecimiento"
    post:
      tags:
      - Establecimientos
      summary: Registra un establecimiento
      description: El código son los cuatro dígitos que asigna SUNAT y aparecen en
        la ficha RUC; no se genera aquí. La casa matriz es 0000.
      operationId: registrar_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionNuevo"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEstablecimiento"
  /api/v1/configuracion/empresas:
    get:
      tags:
      - Empresa
      summary: Empresas que alcanza el usuario
      description: "Las de la cuenta a las que tiene asignación, que son las mismas\
        \ que ofrece el selector de contexto. No las de la cuenta entera: dos usuarios\
        \ de la misma cuenta pueden alcanzar empresas distintas."
      operationId: listar_5
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaEmpresa"
    post:
      tags:
      - Empresa
      summary: Registrar una empresa
      description: "Los datos de SUNAT salen de la atestacion firmada y no del cuerpo\
        \ de la                     peticion: la API no puede consultar el padron\
        \ —su Lambda no tiene salida a                     internet— asi que confia\
        \ en la firma de ondexia.consultas. Responde 409 si                     el\
        \ RUC ya esta registrado, 403 si quien llama no es administrador de la   \
        \                  cuenta, y 400 si la atestacion caduco, si no es valida,\
        \ si el RUC no esta                     habido activo, o si el plan no admite\
        \ otra empresa."
      operationId: registrar_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionAlta"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEmpresa"
  /api/v1/configuracion/empresa/verificacion:
    post:
      tags:
      - Empresa
      summary: Trae del padron los datos que no se pueden editar
      description: "Sirve para dos cosas: refrescar una empresa cuando su razon social\
        \ cambia en SUNAT, y verificar por primera vez una empresa creada en el onboarding,\
        \ cuyos datos los teclearon a mano. Responde 400 si la atestacion no es valida,\
        \ si caduco, o si es de un RUC distinto al de la empresa activa."
      operationId: verificar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionVerificacion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEmpresa"
  /api/v1/almacen/productos:
    get:
      tags:
      - Productos
      summary: Lista o busca productos
      description: "Sin `q`, todo el catálogo. Con `q`, por código o nombre, hasta\
        \ 50."
      operationId: listar_6
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: tipo
        in: query
        required: false
        schema:
          type: string
          enum:
          - BIEN
          - SERVICIO
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaProducto"
    post:
      tags:
      - Productos
      summary: "Registra un producto, disponible en el establecimiento indicado"
      operationId: registrar_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionNuevoProducto"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaProducto"
...
  /api/v1/almacen/almacenes:
    get:
      tags:
      - Almacenes
      summary: "Lista los almacenes, activos e inactivos"
      operationId: listar_7
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RespuestaAlmacen"
    post:
      tags:
      - Almacenes
      summary: Registra un almacén
      description: "El código lo elige el cliente: se teclea en cada movimiento de\
        \ mercadería, así que conviene que sea corto. Se guarda en mayúsculas."
      operationId: registrar_7
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeticionNuevo"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaAlmacen"
...
  /api/v1/configuracion/empresas/{id}:
    get:
      tags:
      - Empresa
      summary: Datos de una empresa del usuario
      description: "Responde 403 si el id no es de una empresa que el usuario alcance,\
        \ y 404 si no existe. Ninguno de los dos mensajes distingue «no es tuya» de\
        \ «no existe» más de lo imprescindible."
      operationId: consultar_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RespuestaEmpresa"
...
components:
  schemas:
    PeticionEdicion:
      required:
      - nombre
      type: object
      properties:
        nombre:
          maxLength: 300
          minLength: 0
          type: string
        direccion:
          maxLength: 300
          minLength: 0
          type: string
        correo:
          maxLength: 200
          minLength: 0
          type: string
        telefono:
          maxLength: 30
          minLength: 0
          type: string
    RespuestaCliente:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipoDocumento:
          type: string
        tipoDocumentoNombre:
          type: string
        numeroDocumento:
          type: string
        nombre:
          type: string
        direccion:
          type: string
        correo:
          type: string
        telefono:
          type: string
        verificadoEn:
          type: string
          format: date-time
        admiteFactura:
          type: boolean
        activo:
          type: boolean
    PeticionEstado:
      required:
      - activo
      type: object
      properties:
        activo:
          type: boolean
    RespuestaCaja:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sucursalId:
          type: string
          format: uuid
        codigo:
          type: string
        nombre:
          type: string
        activa:
          type: boolean
        sesionAbierta:
          $ref: "#/components/schemas/RespuestaSesion"
...
    PeticionPerfil:
      required:
      - apellido
      - nombre
      type: object
      properties:
        nombre:
          maxLength: 150
          minLength: 0
          type: string
        apellido:
          maxLength: 150
          minLength: 0
          type: string
        telefono:
          maxLength: 30
          minLength: 0
          type: string
    DatosDePerfil:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        telefono:
          type: string
...
    RespuestaUsuario:
      type: object
      properties:
        asignacionId:
          type: string
          format: uuid
        usuarioId:
          type: string
          format: uuid
        email:
          type: string
        nombre:
          type: string
        activo:
          type: boolean
        invitado:
          type: boolean
        rolId:
          type: string
          format: uuid
        rolNombre:
          type: string
        sucursalId:
          type: string
          format: uuid
        sucursalNombre:
          type: string
        todosLosEstablecimientos:
          type: boolean
        propietario:
          type: boolean
    RespuestaSerie:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sucursalId:
          type: string
          format: uuid
        tipoDocumento:
          type: string
        tipoDocumentoNombre:
          type: string
        serie:
          type: string
        ultimoNumero:
          type: integer
          format: int64
        siguienteNumero:
          type: string
        activa:
          type: boolean
...
    RespuestaEstablecimiento:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codigo:
          type: string
        nombre:
          type: string
        direccion:
          type: string
        ubigeo:
          type: string
        activa:
          type: boolean
    PeticionEmpresa:
      type: object
      properties:
        nombreComercial:
          maxLength: 300
          minLength: 0
          type: string
        cuentaDetracciones:
          maxLength: 30
          minLength: 0
          pattern: "[0-9]*"
          type: string
        regimenTributario:
          type: string
          enum:
          - NUEVO_RUS
          - OTRO
        permiteVentaSinStock:
          type: boolean
...
    RespuestaRegistro:
      type: object
      properties:
        cuentaId:
          type: string
          format: uuid
...
  securitySchemes:
    jwt:
      type: http
      description: Token de acceso emitido por el grupo de usuarios de Cognito.
      scheme: bearer
      bearerFormat: JWT